ci(actions): skip required CI for non relevant changes
Currently any standalone changes to the readme don't trigger the CI,
which causes the pull request to be blocked because the CI build is
set as required.
To workaround this, use a prep job to check for relevant changes and
trigger a skip job for trivial changes like readme updates.