ci(actions): adjust build trigger check for dependabot
With the check using `github.ref` and dependabot as an actor, the build
is skipped. Previously, only the `push` trigger exists and will trigger
the workflow. As the `pull_request_target` has been added to also allow
pull requests from forks to trigger the workflow, the check needs to be
adjusted.

Related to:
- https://github.com/MediaMarktSaturn/technolinator/issues/378