feat: add option to populate version fields (#704)
This PR introduces a new option to populate the version field in dep
track that is different from the branch name. this option is set to
false by default, so the current behaviour is maintained.

For Releases the tag name is used.
For Pushes and PRs the commit hash is used (I'm open to alternatives!)

This is a very spammy option and not for the faint of heart, as it will
create a new version entry in dependency track each time it's run.

CC: https://github.com/MediaMarktSaturn/technolinator/issues/566